# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ペアPay API",
    "version": "1.0.0",
    "description": "現在実装済みのエンドポイントのみ記載（fetch コードに合わせた）"
  },
  "paths": {
    "/api/callback": {
      "get": {
        "summary": "LINEログインのコールバック",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "カスタムトークンを含むユーザー情報",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": { "type": "string" },
                    "userName": { "type": "string" },
                    "customToken": { "type": "string" }
                  },
                  "required": ["userId", "userName", "customToken"]
                }
              }
            }
          }
        }
      }
    },
    "/api/request/payment": {
      "get": {
        "summary": "すべての支払いリクエストを取得",
        "responses": {
          "200": {
            "description": "支払いリクエスト一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Payment" }
                }
              }
            }
          }
        }
      }
    },
    "/api/webhooks/line/payment": {
      "post": {
        "summary": "新しい支払いリクエストを作成（LINE webhook 経由）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RequestPayment" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "作成成功。 requestId を返す。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "requestId": { "type": "string" }
                  },
                  "required": ["requestId"]
                }
              }
            }
          }
        }
      }
    },
    "/api/request/payment/{requestId}": {
      "get": {
        "summary": "特定の支払いリクエストを取得",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "支払いリクエスト詳細",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Payment" }
              }
            }
          },
          "404": { "description": "リクエストが見つからない" }
        }
      },
      "patch": {
        "summary": "支払いリクエストを承認する／更新する",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["isApproved", "userId"],
                "properties": {
                  "isApproved": { "type": "boolean" },
                  "userId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "承認／更新されたリクエスト",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Payment" }
              }
            }
          },
          "403": { "description": "自分のリクエストを承認しようとした" },
          "409": { "description": "すでに承認済み" }
        }
      }
    }
  },
  "components": {
  "schemas": {
    "RequestPayment": {
      "type": "object",
      "required": ["userId", "userName", "paymentTitle", "paymentCost"],
      "properties": {
        "userId": { "type": "string" },
        "userName": { "type": "string" },
        "paymentTitle": { "type": "string" },
        "paymentCost": { "type": "number" },
        "itemLink": { "type": "string" },
        "paymentMemo": { "type": "string" },
        "category": { "type": "string" }
      }
    },
    "Payment": {
      "type": "object",
      "required": [
        "requestId",
        "userId",
        "userName",
        "paymentTitle",
        "paymentCost",
        "createdAt"
      ],
      "properties": {
        "requestId": { "type": "string" },
        "userId": { "type": "string" },
        "userName": { "type": "string" },
        "paymentTitle": { "type": "string" },
        "paymentCost": { "type": "number" },
        "itemLink": { "type": "string" },
        "paymentMemo": { "type": "string" },
        "category": { "type": "string" },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "approvedAt": {
          "type": "string",
          "format": "date-time"
        },
        "isApproved": { "type": "boolean" }
      }
    }
  }
}

}